requirements.txt
fastapi
uvicorn
gTTS
python-dotenv
qdrant-client
sentence-transformers
google-generativeai
pydantic
requests